requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
idna==3.7
importlib_resources==6.4.0
kafka-python==2.0.2
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
py4j==0.10.9.7
pyparsing==3.1.2
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sseclient==0.0.27
tzdata==2024.1
urllib3==2.2.1
zipp==3.18.1
